[IMPROVES] Add subcategory for Users settings (general, security, blacklist)
